Prevent admin from removing own account

+++ src/main/java/edu/netcracker/project/logistic/controllers/AdminController.java
@@ -11,12 +11,13 @@ import edu.netcracker.project.logistic.service.AdvertisementService;
 
 import edu.netcracker.project.logistic.validation.EmployeeValidator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Collections;
+import java.security.Principal;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
@@ -31,19 +32,21 @@ public class AdminController {
     private AdvertisementService advertisementService;
     private AddressService addressService;
     private EmployeeValidator employeeValidator;
+    private UserDetailsService userDetailsService;
 
 
     @Autowired
     public AdminController(OfficeService officeService, EmployeeService employeeService,
                            RoleService roleService, AdvertisementService advertisementService,
                            AddressService addressService,
-                           EmployeeValidator employeeValidator) {
+                           EmployeeValidator employeeValidator, UserDetailsService userDetailsService) {
         this.officeService = officeService;
         this.employeeService = employeeService;
         this.roleService = roleService;
         this.advertisementService = advertisementService;
         this.addressService = addressService;
         this.employeeValidator = employeeValidator;
+        this.userDetailsService = userDetailsService;
     }
 
     @GetMapping("/crud/advertisement")
@@ -68,11 +71,11 @@ public class AdminController {
     }
 
     @GetMapping("/crud/advertisement/update/{id}")
-    public String showAdvertisementData(@PathVariable long id, Model model){
+    public String showAdvertisementData(@PathVariable long id, Model model) {
         AdvertisementForm advertisementForm = new AdvertisementForm();
 
         Optional<Advertisement> advertisementOptional = advertisementService.findOne(id);
-        if (!advertisementOptional.isPresent()){
+        if (!advertisementOptional.isPresent()) {
             return "redirect:/error/404";
         }
 
@@ -89,10 +92,10 @@ public class AdminController {
 
     @PostMapping("/crud/advertisement/update/{id}")
     public String updateAdvertisement(@PathVariable long id,
-                                      @ModelAttribute(value = "advertisement") AdvertisementForm advertisementForm){
+                                      @ModelAttribute(value = "advertisement") AdvertisementForm advertisementForm) {
 
         Optional<Advertisement> advertisementOptional = advertisementService.findOne(id);
-        if (!advertisementOptional.isPresent()){
+        if (!advertisementOptional.isPresent()) {
             return "redirect:/error/404";
         }
 
@@ -106,13 +109,13 @@ public class AdminController {
     }
 
     @PostMapping("/crud/advertisement/delete/{id}")
-    public String deleteAdvertisement(@PathVariable long id){
+    public String deleteAdvertisement(@PathVariable long id) {
         advertisementService.delete(id);
         return "redirect:/admin/advertisements?delete=success";
     }
 
     @GetMapping("/advertisements")
-    public String getAllAdvertisements(Model model){
+    public String getAllAdvertisements(Model model) {
         model.addAttribute("advertisements", advertisementService.findAll());
         return "/admin/admin_advertisements";
     }
@@ -167,9 +170,17 @@ public class AdminController {
     }
 
     @PostMapping("/crud/employee/{id}/delete")
-    public String deleteEmployee(@PathVariable Long id) {
+    public String deleteEmployee(@PathVariable Long id, Model model, Principal principal) {
+        Optional<Person> opt = employeeService.findOne(principal.getName());
+        if (!opt.isPresent()) {
+            return "error/500";
+        }
+        if (opt.get().getId().equals(id)) {
+            model.addAttribute("message", "Can't remove own admin account");
+            return "error/400";
+        }
         employeeService.delete(id);
-        return "/admin/admin_employees";
+        return "redirect:/admin/employees";
     }
 
     @GetMapping("/crud/employee")
